feat: implement searching and display results
